chore(git): add reference to lsp in ignorefile
Language server protocal binary files, there might be a better place
to put them but as it stands I'm just making  sure they don't get
tracked in the repository.

Packer generates a compiled lua file that is sourced in place of the
plugin list the user edits. I assume this has some optimization?
But ultimatly, it is reproducable so I don't see a reason to commit it.