feat: important scripts! maps, plugs & sets
I held off on adding these until now just to get some other things added to the repository

But with a lot of the configs finally tracked, now comes the real meat of the development environment.

NOTE: These settings were for my windows machine and I have since tested agains Ubuntu and WSL

The new configs are mostly in lua and will replace a fair number of the included scripts in the last few commits, but I wanted to make sure all the important stuff was accounted for before I migrate and merge the new settings.

TL;DR: these scripts are out of date with my current setup and will soon be replaced.