feat: no relative number in insert mode
This script used to be at the bottom of my init/vimrc file but I have moved it into it's own plugin file.

Some considerations with it are that it currently does not respect whether the user had previously enabled gutter numbers, relative or not.

So inside the function I would like to add a check for if the user had previously enabled either one.