feat: add module to control builtin plugins
This file simply manages the configuration of preinstalled plugins for
vim/neovim such as netrw. It defines a number of globals but primarly it
controls whether these plugins should be initalized on startup.
It does so by setting the `loaded_{plugin}` global variable which the
plugin uses to determine if it has already been initialized.