feat(plugins): include native-lsp plugin definitions
Native lsp has some benefits and some drawbacks over coc at the moment,
Most of which relate to performance and ease of use respectively.
The current configs still have an issue with attaching the buffer
properly and needs some tweaking, but with a few docker tests it seems
as if it handled clean installation pretty well thanks to lsp-installer.