feat(lsp): add configs for sh/bash language-server
Whether or not this works with zsh is undetermined from my testing, it
doesn't appear as if the repo makes any claims to support it but at
least some of the syntax would work. However, explicitly adding the
`zsh` extension as a file type did not seem to trigger the server
attachment.