feat(lsp): add basic efm-langserver configs
EFM is a general purpose Language Server that depends on golang being
installed on the host system.

Testing has not be done but the goal was to utilize this in place of the
lua based formatting.nvim plugin to handle auto-code formatting/indents.
Driving this with a general lsp could have some benefits for wider
compatibility between different languages. However, the need for go is
not completely ideal.