feat: add whichkey mapping tables
I prefer that as many custom keybindings are defined in the same file.
To that end, these settings are specific to the lua reimplementation of
whichkey. The script will check if it exists before loading the mappings
so it is not a hard dependency. But I'd be lost without some of these
settings.