fix(plugins): fix packer config module loader
This fix moves the config table and require statement into an anonymous
function that is evaluated after the bootstrap function and plugin list.

Packer config is stored in a separate module from the plugin list.
When attempting to bootstrap packer from a fresh nvim: this file is
loaded and attempts to call `packer.util` before the bootstrap function
is finished.