refactor(dap): move dap configs to plugin module
The dap plugins is references simply with `require('dap')`;
this is a problem because lua is attempting to look in `lua/dap`
for user configuration.

To avoid this issue, it's best to move the dap configs to a new
namespace. Such as, `lua/plugins/dap`.

I was originally against it due to it sitting nicely with `lua/lsp`,
but the rational doesn't work if you consider that neovim lsp is
builtin, and dap is not!