refactor: remove python-dap from required plugins
In the packer plugin list under the base dap reference:
python-dap definitions were listed as a required or dependent plugin.

Given that packer has provided me with a 'failed to install/update'
error for this plugin a number of time. As well as the fact that I can
easily roll my own adapter and configuration tables. It seems best to
remove this dependency.