feat(dap): config loader for dap module
The config function will now search `plugins/dap` for lua scripts that
corresponding to the language they belong.

If python exists in the language list, and `plugins/dap/python.lua`
That module will be loaded to initialize the adapters and configurations.

The language list above simply acts as a feature flag.
to prevent language that are not in the list, i.e. untested:
from being initialized.