fix(plug): explicit 'no' signcolumn in toggleterm
Treesitter-viml has a bug where `set signcolumn=no` breaks the following
lines highlighting, leading one to believe the statement is incorrect.
While the same does not occur for `set signcolumn=`, which is valid in
the Ex command mini buffer but is not valid in vimscript files.

Additionally, the runtime plugin/terminal can handle this setting but
for anyone who uses this configuration file for toggleterm, this is a
helpful thing to set explicitly for the plugin file type.

Think of it as added security for a clean terminal interface!