fix: explicit value for all open ended sets
Any `set` or `setlocal` that had an empty value such as `set signcolumn=`,
now has a value assigned to it.

The issue lies with a bug in treesitter-viml leading me to beleive that
`signcolumn=no` and similar settings variables would not work properly
due to broken syntax highlighting.

However, this is not the case; not having the value assigned after
the equal sign results in error messages.
Worse than slightly ugly syntax issues.