build: add runtime path script for testing vim
This script loads vim with the config and runtime paths set to the
current directory. All paths will point to a local xdg home layout.
A symlink is created to the current directory and `./.config/nvim`.

This way we can have a special version of neo/vim for testing
configuration without causing a mess for the rest of our environment.