refactor: overhaul lsp_installer settings in lsp.init
- Fix missing lsp_installer module reference
- Add auto install from default list of servers
- Set default and conditional settings for language servers

refactor: rename all lsp configs to setup

Setup seems to be the standard and this is either confusing or can be
seen as a wrapper for the standard call to a plugins setup function.

`require('telescope').setup({ opts = opts })`

But your table is shipped off to lsp.init in order to be processed by
the lsp_installer list.

refactor: lsp installer extend options table for defaults

Set a default value in the `options` table, when using the table extend
method with the 'force' flag: the table keys and values should be
concatenated from right to left.