refactor: remove pcall plugin checks in mappings
Initially the idea was to not set the mappings if the particular plugin
could not be located. However, there are no alternative mappings for
other plugins on the same whichkey sequence.

Any resulting errors from a missing plugin will only be presented when
attempting to use the particular key sequence and will not prevent
neovim from loading as expected.