fix: apply missing feat plugin, spellsitter config
During the plugin config refactor I missed a development branch plugin
I was testing that ended up getting committed: 'spellsitter'.
This is a treesitter extension that allows for `set spell` to work only
on comment text, as opposed to displaying misspellings throughout the
code syntax.

With the plugin already being installed, we just need to initialize the
setup function and ensure that `set spell` is globally enabled.