refactor: replace fidget with lualine_lsp_progress
Backpedaling on fidget plugin due to display issues.
Preferred to have fidget display globally at bottom right hand
side of editor, but this doesn't play well with bottom drawer terminal.
As it will display on top of the terminal when it is open.

Instead, we can use lualine_lsp_progress to display in the active
status line. Preserving the scope of the lsp actions and keeping a
cleaner look.