fix: sumneko lsp long loads and environment prompt
Fixes: #5

Sumneko supports lua LOVE as a third party environment, but currently is
more of a nuisance when opening any sort of lua file for neovim config.
This disables the 3rd party check, eliminating the annoying prompt.

On top of that, initialization time for the sumneko language server was
taking a day and a half to fully load. This seems to be caused by making
the lsp aware of all runtime files which includes packer plugin files.

This could be helpful for developing and debugging plugins but basically
renders the language server useless for up to 10 minutes on startup.