fix: define sumneko default library/workspace
Set paths for sumeko to analyze at startup.
Limit the number of files to load to keep initalization of the server optimized.

You can add a path to this table to get access to references/declarations.
Useful for developing or contributing to nvim-lua plugins by adding the
following to your library:

```
-- Add packer to library for plugin definitions
[vim.fn.stdpath('data') .. '/site/pack/packer/start/*'] = true,
[vim.fn.stdpath('data') .. '/site/pack/packer/opt/*'] = true,
```