fix: override lualine theme vscode: inactive highlights
refactor: invert color for b-c components for softer appearance.

fix: local patch for improper highlight on inactive components

This themes is setup in such a way where inactive components are
rendered with white backgrounds for things such as buffers/windows/tabs.

Likely not noticed if the users don't utilize lualine as their tabline.