refactor: improve colors module helpers
This module was created to handled various issues with applying
highlights after color schemes being loaded. With the way neovim/lua
plugins have developed, it's become more common for the themes setup
function to expose a highlight override table, making that aspect of
this module less useful but not all color schemes provide this
functionality, so this will remain.

More changes:

- Ensure colors module is reloaded to reapply theme.setup function
- Improve get_highlight function to better handle extending highlights
- Add function to check if color scheme exists in runtimepath
- Add function to extend higroups, since nvim_set_hl doesn't handle this
- Add type hints and doc comments for helper functions
- Rename setup helper to load, as it handles init and reload
- Cleanup theme loader function and add asset / fail on error
- Rename color/init level highlights to global_overrides for clarity
- Remove extraneous global_overrides entries
- Rename function keywords to `M.name = function`
- Fix remove `ui` and `clear` from color scheme module highlight tables