fix: disable tabout, causing indent issues
This plugin is nice to have but causes a few more problems than it
solves. While in insert mode and a pair is present, such as `[({"''"})]`
the tab key will attempt to "tabout" at inappropriate times which
prevent proper indenting of the desired line.

If and until this can be fixed the plugin spec will be present with the
flag `enabled = false` to prevent it from being loaded.

The plugin will likely be removed if a good solution cannot be found.