fix(core): asterisk hl cursor word shouldn't jump
Overwritten keymap didn't function properly due to incorrect remap.

By default, in normal mode `*` will highlight the word under the cursor,
but it will do so by jumping to the next occurrence of the word within
the current buffer, which is pretty annoying.

Previous attempt to override this extra jump behaviour used `<C-p>` as
an attempt to jump to previous hlsearch. The proper keymap seems to be
shift+n `<S-n>` or simply uppercase `N`.