refactor: replace guess-indent with vim-sleuth
Guess-indent is a nice plugin but it's algorithm might be too dumb in
it's current state for my liking. I will often read a buffer with an
expected indent size, set from options < after/ftplugin < editorconfig.
But the plugin has struggled to infer when cloned projects have a
different indentation size.

Vim-sleuth is much older and while it's unfortunate I have to lean on a
vimscript plugin, it is a proven and battle tested plugin in it's own
right.