fix: prevent signature help from overlapping cursor
Use defaults for max width for floating window and above cursor affinity
for lsp_signature plugin.

Width limit was for stylistic reasons but using the defaults is fine.
Cursor relative positioning was disabled with the assumption that it
would be better for readability, however there are cases when the cursor
is on the botright of the editor window, the cursor and text with be
overlapped by the floating window.