fix(plug): revert co-opt dashboard group for filetype overrides
Define a custom augroup for dashboard filetypes autocmd to apply buffer
local options.

Previous commit removed the custom augroup in favor of using the
predefined `dashboard` augroup. I should have expected that doing so
would cause multiple autocmds to be added to the group if `opts.clear =
true` was not set for the filetype autocmd.