feat(command): add basic command `R` to reload lua modules
The command can be used on the active lua buffer eg: `init.lua`. Or an
argument can be passed for a specific package name eg: `telescope`.

It will either use plenary or `package.loaded = nil` to clear the
package cache and then require the module from a fresh state. You can
use the completion menu to list all available modules that are listed in
the `package.loaded` table.