fix(ft): prevent log files from being edited
Log files are meant for reading! An inconsequential issue I often have
is accidentally writing characters into log files. To prevent this we
can simply set all log filetypes to `nomodifiable` and set their buftype
to `nowrite`.

Should you have need to modify the content of a log file, I ask... Why?
Either way, I recommend creating a new buffer and editing the data there
to save it as a note. Or, you can set the buftype and modifiable state
for the current buffer manually.