feat(zenmode): make zenmode more context aware
refactor: define on_open outside of setup

feat: add buftype and filetype exclusion rules

This simply makes zenmode act as a fullscreen zoom window instead of
restricting the size to the current textwidth value.

feat: dynamically resize zenmode

Takes into account the size of the current buffers textwidth, the size
of the gutter and number line and the spacing of the colorcolumn to
dynamically set the width of zenmode.