feat(toggleterm): add custom terminal layouts
These terminals simply have custom settings for how they are displayed
IE. a floating window at the bottom of the screen to mimic a horizontal
split, similar to how telescope pickers are displayed. This can be
helpful to prevent windows from being pushed around but comes with some
caveats. Mostly this is in the form of buggy affects when focus is
shifted as a new window is created.

The bottom floating window can also be used within zenmode
since it does not allow for splits to be created when zoomed in.

Additionally there are other layouts such as fullscreen_float which
basically makes it seem as if neovim is not even open and you're back in
the main terminal. Or, the tab_float which take into account where the
buffer and statusline are located to mimic how a new tab would look.