fix(ts): remove powershell parser config
I have gotten a lot of errors with this due to ABI version mismatches. It seems this parser has been abandoned/archived. There is an official one on the powershell github organization but it is under-developed at this time.

To prevent further errors with powershell scripts, this will remain removed until further notice.