fix(auto-session): improve session save/delete notify
Default logging for plugin events are rather verbose and can be annoying
if the relevant path string is too long, resulting in the need to "Press
ENTER to continue" prompts.

This sets logging back to `error` to suppress default logging output to
the echo area. Following this change includes a basic wrapper that
injects a custom output message for `SessionSave` and `SessionDelete`
user commands.