fix(cmp): improve cmp sorting comparators
Add `recently_used` and `locality` comparators to get completion items
to be more context aware. Should the user type the same with multiple
times, like a class name followed by any of it's methods, the class will
be ideally be the first item in the completion menu due to repeated use
or buffer proximity.