fix(cmp): add trigger characters for emmet_ls completion
Some snippet definitions or language servers such as emmet_ls can have
completion strings that start with a bang "!" or other non-standard
characters like pound "#".

Since these are not defined as default trigger characters the snippets
that utilize them cannot trigger the completion menu to be displayed
with a single character. In order to get this functionality we need to
append these characters to the list of trigger characters.