fix(cmp): remove duplicate items in completion menu
NOTE: This is currently being tested and could be reverted if the
results are unfavorable.

Setting the format value `vim_item.dup = nil` prevents items that share
the same name from being displayed. This could lead to potential issues
where a desired item will ever be displayed. An example of this is
perhaps custom snippets that share a prefix value. A snippet plugin such
as friendly-snippets might not allow custom snippets from being
presented.