feat(cmp): reinstall source: `hrsh7th/cmp-nvim-lua`
Source was removed in favor of using `folke/neodev.nvim` and lua_ls to
provide type information for neovim runtime definitions. Unfortunately
this cannot be utilized with the cmdline context, leaving items such as
`vim.api.*` to be marked as variables instead of functions.

By reintroducing the plugin and enabling the source we can get proper
type information for neovim commands.