refactor(cmp): improved completion source chaining
Defining sources using multiple tables in a chain as the arguments for
`cmp.config.source` will prevent sources that are lower in the chain for
being displayed if there are still valid items from the previous chains.
This way, the completion list is not polluted with too many irrelevant
items.

This configuration will need some more hands-on testing to see if any
further tweaks are in order to improve user convenience. But logically,
we simply don't want buffer or spell items to get in the way of LSP
types or snippets when they are available.

This also removes the configuration for markdown filetypes, as the
global source definitions are now practically identical. Saving the need
to define multiple filetypes alongside markdown for a more robust
"documentation" context. IE. Including filetypes such as plaintext,
LaTeX, restructured text and/or org/neorg documents.