fix(cmp): remove path keyword_length
Since adding the performance option to limit the number of completion
items to ~25 entries I haven't experience any issues with the `/tmp/`
directory crashing neovim/wsl as cmp attempted to spawn the completion
window with god know how many potential paths...

Ontop of that, there is a bug with the keyword_length setting where it
will still attempt to complete the path when your cursor is located
in front of a symbol such as a forward slash. Eg. `/path/and/cursor/|`
will still display results until the first character is typed, then the
keyword_length will kick in. Making path completion more of an annoyance
without actually fixing anything since this was originally a hack or
bandage approach.