fix(lualine): prevent cursor from jumping to prev snippet
There are times when pressing the `<tab>` key after expanding a snippet,
the cursor will jump to a previous snippets node somewhere else in the
document. This is really annoying if you indented to insert indentation
and not jump 20 lines up to a "now" irrelevant block of code.

This attempts to fix this by setting a region check event for luasnip
that is disabled by default.

Documentation:

`region_check_events`: Events on which to leave the current snippet if
the cursor is outside its' 'region'. Disabled by default, 'CursorMoved',
'CursorHold' or 'InsertEnter' seem reasonable.