fix(dashboard): increase header top padding
Simple fix to add line padding above the header artwork, giving it the
same amount of padding between the center and footer section.

Having a more dynamic approach would be preferable but would introduce
some overhead. There is an open PR for this functionality in the plugins
upstream repo but this hack is perfectly fine as I don't want to
vertically center the WHOLE dashboard.