fix(dashboard): set header version via IIFE function
Refactors the header section for displaying the current neovim version.
Using an IFFE or iffy function allows the logic to be evaluated and
remain inline with the header table.

Replace the pre-release text `dev` with `nightly`.

Micro-optimize loading of the plugin options by using a variable for
`vim.version` to prevent duplicate function calls.