fix(ts): remove python from ts-indent: disabled_filetypes
Improvement to the python parser has made indentation more bearable.

The situation was fairly bad for previous versions of the parser causing
wildly inaccurate indentations which really wasn't fun to work with.
Referencing other peoples dotfiles, it seemed like a common solution was
to just disable python from the extension settings.

Alternatively, we could have install a plugin such as autopep8-indent
but I don't like this idea, as it's yet another plugin; the plugin is
older and ultimately a proper solution in treesitter is greatly
preferable.