perf(telescope): prevent loading for extension deps
The plugins auto-session and project_nvim support telescope integration
but there doesn't seem to be a cleaner way of loading the extension
within their respect config/specification.

A compromise is to pcall load the extensions inside of telescope's
config callback, which allows the core plugins to be loaded while
allowing telescope loading to be defered until the user requests access
via the `Telescope` user command, saving about 20ms on startup time.