fix(npairs): disable paren completion for powershell functions
Many cmdlets don't use parenthesis for function invocation so this is
more of a nuisance with general powershell syntax. Though, there are
aspects of the language that do use parenthesis in certain contexts. So
this change isn't foolproof, it's just a better compromise than adding
pairs when completing ANY function type.