feat(autocmd): replace lazyredraw opt w/ autocmd for large buffers
Every time the user switches the active buffer, it will run the autocmd
and check if the number of lines hits a certain threshold. This is
currently set to 1000 lines but the value can be further tweaked after
some testing.

This should help for running macros and regex replace commands without
any potential negatives of keeping lazyredraw perpetually enabled.