fix(lualine): change gitsigns dict access to remove linter warning
Since the value might not be defined, lua_ls will complain that the
value accessed might not be defined, to prevent this warning from being
displayed we used luadoc comments to disable the warning but we can also
use square bracket notation to access the buffer table and pull out the
gitsigns dictionary by it's key name. Which will return nil if it is not
available.