feat(lualine): explicitly set tabline component colors
Use `lualine_c` for active/inactive windows and tabs. The colors should
default to white for active and gray for inactive.

The first preferable choice was to use `use_mode_colors = true` but this
setting has a bug where the highlights are cleared in the tabs
component. This bug occurs because if the statusline sections `b` or `x`
use components with complex highlights like `filetype` or `diagnostics`,
highlights are not properly set for the `tabs` component within the
tabline configuration.