fix(dashboard): change icons due to shifted cursor position
The previous nerdfont icons have a longer escaped hex value and when
measuring the string length of a single icon, they equate to a value of
greater than one. Which causes issues with dashboard.nvim logic for
padding and/or where the display the cursor for the active selection.

The new cursors have a hex length of 4 and don't suffer from the same
issue while still representing the category of their respective
commands.