fix(lualine): revert use of `list_contains` until a later version
Distro hopping in a virtual machine and the AUR package for
neovim-nightly is out-of-date, leading to a number of error messages
from the lualine component: lsp_client. This is due to the use of
`vim.list_contains` not being available in the currently available
version. I could build nightly from source but I don't have the time for
that.

As a means of quickly suppressing this error, we will return to the use
of `vim.tbl_contains` but keep the use of `vim.list_extend` to avoid
previous bugs where clients were not populated properly.

This also includes a TODO comment regarding a deprecation warning from
lua_ls to replace `vim.lsp.get_active_clients` with
`vim.lsp.get_clients`.