fix: lint warning via luadoc type cast for fold helper function
Function `vim.fn.getline` returns type `string|string[]`, which throws
off lua_ls linting when the return value is being passed to a function
that expects only a string.

Ideally, it might be better to properly handle the possibility of a
table value via type checking but this bloats the helper function when
`getline` will realistically only return a string since it is being
supplied a single argument.

With that in mind, the warning can also be suppressed via diagnostic
ignore hints or by inserting casting hints for the respective variables.