fix(neogit): force commit window to use filetype: gitcommit
Neogit uses it's own filetype for commit buffers. This has been
uncommitted in my config for awhile and I'm unsure of neogit still uses
this custom filetype `NeogitCommitMessage`, but I'd rather it use the
default filetype regardless.

With that said, this could be reverted in the future after some more
testing, but my current system for testing worktrees doesn't work with
the neogit commit system.