fix(lsp): remove unnecessary signature_help autocmd
The documentation claims that the plugin can be setup via the plugins
setup function or alternatively via the client on_attach function. I
believe it to be cleaner to let this be initialized via the plugin spec
instead of littering the local on_attach or hacking a LspAttach autocmd
to do the job.