feat(lsp): add plugin lightbulb.nvim for code actions
I want a similar way of displaying available code actions for the
current line the same way an IDE would, and this plugin does the trick
nicely.

However, given that the builtin diagnostic hints for code actions, it
would be nice to not need this plugin and simply filter hint diagnostics
that are in-fact "code actions" and only display them for the current
line context.