fix: prevent process exited prompt in toggleterm buffers
For regular terminal buffers that are spawned with the `terminal`
command, this does not occur unless an argument is provided. allowing
the user to effectively read the output of the supplied command.

`[Process exited 0]`

ToggleTerm buffers that are restored from vim sessions don't properly
closing, displaying the aforementioned exit message. This should never
happen with restored ToggleTerm buffers. Likely, the buffer is not
persisting some condition/event via sessions. By adding an autocmd on
plugin load, we can restore the behaviour after the session has been
reloaded.