fix: relocate foldtext func to avoid global vars
This linting rule was previously disabled within luacheck. However,
since moving to selene, global variables are also not recommended by
default and I believe this is a simple change that is correctly more
idiomatic.

This change requires the foldtext function to be moved to another module
for ease of access. This is due to the fact that that vim foldtext
option requires a string value that will be evaluated:

`opts.foldtext = "lua:require("module").custom_foldtext()"`.