fix: remove formatoption to allow comments to autowrap
Previous autocmd for all filetypes removes the formatoptions `cro` to
prevent auto formatting for comment lines, such as autowrap and auto
inserting a comment leader on the creation of a newline.

This removes the negative assignment of the formatoption `c` for the
autocmd to re-enable the comment auto-wrapping functionality for all
filetypes, while still preventing the auto insertion of comment leaders.