fix(cmp): set cmp to use PmenuSel for entry selection
This is more inline with how default neovim works and is easier to
visually parse than using the CursorLine highlight depending on ones
colorscheme. My personal preference is to darken the CursorLine itself
if it is even on to begin with. Though, as a result this is much more
difficult to see in lighter environments and/or on darker screens.