fix(lualine): illegal character in lualine component: lsp_info
For some reason, defining the `max_size` with the current configuration
causes an error within lualine stating the use of an illegal character.
I tried opening an issue with the repository maintainer but my issue was
deemed insufficient for reproduction. Though it appears to still be
present as an edge case error.

While it would be nice to define the `max_size` value to prevent
messages from overflowing the whole statusline, this bug needs to
squashed first because it is extremely invasive when trying to do work.