feat(lualine): use the modeline color for tabline
Displays the active window/tab using the same color as the mode
component. This has been a back and forth over a number of commits when
tweaking lualine, but the use of the `use_mode_colors` option means we
no longer need to patch the component for the same functionality.

I will likely keep this for the time being unless I see a need for
displaying status information as color, such as changing the color to
red if the active buffer has not yet been written.