feat: add util wrapper register_keymaps
Allows keymaps to be applied by using a list of keymaps with two tables.
One for options to apply to all keymaps and the second for the keymaps.

The benefit of using this wrapper is that each keymap can also accept
options to be merged with the first set of defaults. This way you can
set the whole list of keymaps to be `silent=true` but each keymap can
have either own set of options to overwrite the supplied defaults.

Example:

```lua
-- Usage of the util function:
local default_opts = { buffer = 0, silent = true }
require("config.util").register_keymaps(default_opts, {
  { "n", "<leader>a", callback, { silent = false, desc = "action: a" } }
  { "n", "<leader>b", callback, { expr   = true,  desc = "action: b" } }
  { "n", "<leader>b", callback, { remap  = true,  desc = "action: c" } }
})

-- Resulting table of keymaps:
-- { "n", "<leader>a", callback, { buffer = 0, silent = false,               desc = "action: a" } }
-- { "n", "<leader>b", callback, { buffer = 0, silent = true,  expr  = true, desc = "action: b" } }
-- { "n", "<leader>b", callback, { buffer = 0, silent = true,  remap = true, desc = "action: c" } }
```