fix: pin none-ls to commit before deprecation warning
I am in the process of updating the plugins list for my neovim config.

It seems a number of builtins are being deprecated in favor of newer
alternatives or language servers that include the same functionality.

On its face, I don't exactly agree with this change but I lack solid
understanding and can't be fucked to deal with it right now. So, I'm
pinning the version to just before the deprecation warning were
introduced.

This can be reverted once a preferred resolution can be determined.
For more information please see the relevant issue:

https://github.com/nvimtools/none-ls.nvim/issues/58