fix(plugin)!: disable problematic plugins if nvim >= v0.10
Two plugins have not been updated to accommodate the changes introduced
in the official release of neovim v0.10.

The plugins use `vim.fn.has` as a condition within their lazy.nvim spec.
So, if the function or expression evaluates to `false`, the plugin will
not be enabled when neovim starts.

These plugins are:

- [ahmedkhalf/project.nvim](https://github.com/ahmedkhalf/project.nvim)
- [kosayoda/nvim-lightbulb](https://github.com/kosayoda/nvim-lightbulb)