fix(lualine): invalid expr for lsp_info component default condition
The condition is supposed to check if there are any active clients
attached to the current buffer. This is a simply integer value check to
see if the list of clients is non-zero. However, this expression missing
a pound symbol `#` to inform lua to return the length of the list