fix(mason)!: remove gopls from ensure_installed
If the go binary is not installed or does not meet a minimal version for
gopls to be installed, mason will present an error/warning message that
the server could not be installed.

To prevent this error message, we will simply remove this entry. The
user can still install the language server manually through the mason
interface.