fix(nls)!: update none-ls and replace deprecated builtins
BREAKING CHANGES:

- Remove support for lua_format

In a previous commit, the none-ls plugin was frozen to prevent breakages
due to a major deprecation that removed or relocated several builtins to
another repository (nvimtools/none-ls-extras.nvim).

This update includes:

- Updating none-ls
- Replacing deprecated builtins
- Removing support for lua_format
- Replacing ruff builtins in favor of ruff_lsp
- Adding ruff_lsp to mason ensure_installed

Unfortunately, this change introduces an additional dependency for
luacheck and removes support for lua_check, as these tools are not
included in the extras repository. While I have transitioned to using
selene and stylua, which are popular within the Neovim community, not
all projects have. This makes it challenging to work within the specific
parameters of certain projects.

Additionally, I have not done extensive testing, but I believe there
might be an issue with code actions being published for things such as
"organizeImports." I'm also uncertain if the document formatting filter
logic will allow Python projects configured to use "black" to continue
doing so without ruff trying to take over. Hopefully, that is not the
case, but bear in mind this could be a regression.