feat!: replace nvim-tree with oil.nvim
There are a lot of nice features in nvim-tree but over time I find
myself using the file tree less and less as I am more accustomed to
fuzzy finding. But for the odd cases where I need to navigate/modify the
directory structure, oil just seems far more intuitive to work with!