feat(colors): make colorscheme persistent
Added a mechanism to remember and reload the last used colorscheme on
Neovim startup. This change ensures the selected colorscheme persists
across sessions by storing the colorscheme name in a global variable.

The autocmd for 'VimEnter' now loads the last used colorscheme, while
the 'ColorScheme' autocmd updates this variable whenever a new
colorscheme is applied.