feat: add user command to redirect output to a new buffer
Added the `Redir` user command, which redirects the output of a given
command to a new buffer. For example, `Redir lua = require("vim.lsp")`
creates a modifiable, searchable buffer with the command's output.
This is useful when the `messages` command falls short due to its lack
of searchability and limited accessibility for larger outputs.