fix(diffview): add missing commands for lazy loading
Added `DiffviewFileHistory` and `DiffviewLog` to the plugin spec to
expose missing commands for lazy loading. These commands were previously
unavailable until the plugin was loaded, but are now accessible and will
trigger the plugin to load when called.

Commands  like `DiffviewToggleFiles` and `DiffviewClose` do nothing
until the plugin is loaded, so they don't require lazy loading.