fix(mason): pin mason to v1 due to dropped support for lsp handlers
Due to dropped support for handlers in mason/mason-lspconfig v2, we are
pinning our config to pin both plugins to v1 until we can figure out how
best to implement this behaviour in the new version.

Previously, we did roll our own handler methods, but testing is needed
to determine if its sufficient to simply pull those back into the
project.